% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcstvvecm.R
\name{fcstvvecm}
\alias{fcstvvecm}
\title{Forecasting of I(1) CS-TV-VECMs Under Restrictions}
\usage{
fcstvvecm(tsrs, objt, hriz = 1)
}
\arguments{
\item{tsrs}{Time series to forecast.}

\item{objt}{Result of a call to \code{cstvvecm}.}

\item{hriz}{Positive integer indicating the forecasting horizon.

Default is \code{1}.}
}
\value{
Returns the time series to forecast
extended by the forecasted values.
}
\description{
Forecasts an I(1) CS-TV-VECM under restrictions.
}
\references{
Veldhuis, S. and Wagner, M. (2026). Restrictions in a State Space Approach
to Time-Varying Vector Error Correction Models: Modelling Instabilities
in Long-Run Money Demand. Mimeo.
}
