% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frstvvecm.R
\name{frstvvecm}
\alias{frstvvecm}
\title{Forecasting of I(1) RS-TV-VECMs Under Restrictions}
\usage{
frstvvecm(tsrs, objt, hriz = 1)
}
\arguments{
\item{tsrs}{Time series to forecast.}

\item{objt}{Result of a call to \code{rstvvecm}.}

\item{hriz}{Positive integer indicating the forecasting horizon.

Default is \code{1}.}
}
\value{
Returns the time series to forecast
extended by the forecasted values.
}
\description{
Forecasts an I(1) RS-TV-VECM under restrictions.
}
