% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvecm.R
\name{fvecm}
\alias{fvecm}
\title{Forecasting of VECM for I(1) Processes Under Restrictions}
\usage{
fvecm(tsrs, objt, hriz = 1)
}
\arguments{
\item{tsrs}{Time series to forecast.}

\item{objt}{Result of a call to \code{vecm}.}

\item{hriz}{Positive integer indicating the forecasting horizon.

Default is \code{1}.}
}
\value{
Returns the time series to forecast
extended by the forecasted values.
}
\description{
Forecasts a VECM for I(1) processes under restrictions.
}
\references{
Johansen, S. (1995). Likelihood-Based Inference in Cointegrated Vector
Autoregressive Models. Oxford University Press, Oxford.
}
