% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/johi1test.R
\name{johi1test}
\alias{johi1test}
\title{Johansen (1995) Test}
\usage{
johi1test(
  sigl,
  tsrs,
  ordr = 1,
  dpow = 0,
  excl = FALSE,
  smpl = 1000,
  simu = 10000,
  tolr = .Machine$double.eps
)
}
\arguments{
\item{sigl}{Significance level.}

\item{tsrs}{(Multivariate) time series to be tested for the cointegrating
rank.}

\item{ordr}{Order of the VECM. Equal to one plus the number of lags.

Default is \code{1}.}

\item{dpow}{Highest power of polynomial time trends (e.g., -1:
no deterministic components, 0: constant only, 1: constant and linear trend)

Default is \code{0}.}

\item{excl}{Logical value indicating whether the highest power
of polynomial time trends is restricted to the long-run
relationship.

Default is \code{FALSE}.}

\item{smpl}{Sample size to consider in case critical values
have to be simulated.}

\item{simu}{Number of repetitions in case critical values have to
be simulated.}

\item{tolr}{Tolerance limit for lookup tables of critical values.}
}
\value{
A list containing:
\itemize{
\item \code{stat}: Test statistics
\item \code{crit}: Critical values
\item \code{rjct}: Logical value indicating the rejection of null hypotheses
}
}
\description{
Performs a Johansen (1995) test
}
\references{
Johansen, S. (1995). Likelihood-Based Inference in Cointegrated Vector
Autoregressive Models. Oxford University Press, Oxford.
}
