% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jvecm.R
\name{jvecm}
\alias{jvecm}
\title{Estimation of VECM for I(1) Processes Using Johansen (1995) Methodology}
\usage{
jvecm(tsrs, cirk, ordr = 1, dpow = 0, excl = FALSE)
}
\arguments{
\item{tsrs}{Matrix of values for multivariate I(1) process.}

\item{cirk}{A positive integer specifying the cointegrating rank.}

\item{ordr}{Order of the VECM. Equal to one plus the number of lags.

Default is \code{1}.}

\item{dpow}{Highest power of polynomial time trends (e.g., -1:
no deterministic components, 0: constant only, 1: constant and linear trend)

Default is \code{0}.}

\item{excl}{Logical value indicating whether the highest power
of polynomial time trends is restricted to the long-run
relationship.

Default is \code{FALSE}.}
}
\value{
Returns a list of the following values:
\itemize{
\item \code{cint}: Estimated long-run coefficient matrix
\item \code{ldng}: Estimated loading/adjument coefficient matrix
\item \code{rvar}: Estimated variance-covariance matrix of the error term
\item \code{rrkc}: Matrix product \code{ldng} * \code{cint}'
\item \code{agmt}: Estimated coefficient matrix of the short-run dynamics
\item \code{cior}: Orthogonal complement of \code{cint}
\item \code{ldor}: Orthogonal complement of \code{ldng}
}
}
\description{
Estimates a VECM for I(1) processes using the Johansen (1995) methodology.
}
\references{
Johansen, S. (1995). Likelihood-Based Inference in Cointegrated Vector
Autoregressive Models. Oxford University Press, Oxford.
}
