% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupjo95qntl.R
\name{lookupjo95qntl}
\alias{lookupjo95qntl}
\title{Lookup critical values for Johansen (1995)}
\usage{
lookupjo95qntl(prob, dpow, hypo, excl, tolr = .Machine$double.eps)
}
\description{
Uses an internal lookup table of Johansen (1995) critical values
stored in \code{sysdata.rda}.
}
\keyword{internal}
