% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrr.R
\name{rrr}
\alias{rrr}
\title{Estimation of a Reduced Rank Regressions Under Restrictions}
\usage{
rrr(
  yvls,
  xvls,
  uvls,
  rrst,
  lrst,
  arst = NULL,
  fish = FALSE,
  tolr = 100 * sqrt(.Machine$double.eps),
  mxit = Inf
)
}
\arguments{
\item{yvls}{Matrix of dependent variables.}

\item{xvls}{Matrix of (reduced rank coefficients related) regressors.}

\item{uvls}{Matrix of (not reduced rank coefficients related) regressors.}

\item{rrst}{Specification of the restrictions on the row space
parameter matrix. Either \code{NULL} or a list of length 3. If list of
length 3, then first element restriction matrix, second element restriction
vector, third element vector of initial values for the free parameters.}

\item{lrst}{Specification of the restrictions on the column space
parameter matrix. Either \code{NULL} or a list. Must have length 4 if
\code{rrst} is \code{NULL}, and length 3 otherwise.
\itemize{
\item If list of length 4, then first element restriction matrix,
second element restriction vector, third element vector of initial values
for the free parameters, and fourth element initial value
for the row space parameter matrix.
\item If list of length 3, then first element restriction matrix,
second element restriction vector, third element vector of initial values
for the free parameters.
}}

\item{arst}{Specification of the restrictions on the coefficients of
the input vectors in measurement equation. If not \code{NULL}, \code{NULL} is returned.

Default is \code{NULL}.}

\item{fish}{Should fisher information matrix be returned?

Default is \code{FALSE}.}

\item{tolr}{Tolerance limit for optimization.

Default is 100 * sqrt(.Machine$double.eps).}

\item{mxit}{Maximum number of iterations.

Default is Inf.}
}
\value{
Returns \code{NULL} if \code{arst} is not \code{NULL}, otherwise
a list of the following values:
\itemize{
\item \code{lfct}: Estimated column space parameter matrix
\item \code{rfct}: Estimated rows space parameter matrix
\item \code{lfun}: Estimated unrestricted column space parameter vector
\item \code{rfun}: Estimated unrestricted row space parameter vector
\item \code{evar}: Estimated variance-covariance matrix of the error term
\item \code{augm}: Estimated coefficient matrix of the short-run dynamics
\item \code{lrco}: Cointegrating rank
\item \code{fish}: Fisher information marix (if argument \code{fish} is set to
\code{TRUE}, otherwise \code{NULL})
\item \code{unvr}: Inverse of the Fisher information matrix (if argument
\code{fish} is set to \code{TRUE}, otherwise \code{NULL})
\item \code{idnt}: Logical indicating whether identification is achieved
(if argument \code{fish} is set to \code{TRUE}, otherwise \code{NULL})
\item \code{nlik}: Negative of the log-likelihood (apart from constants) evaluated
at the optimal solutions found
}
}
\description{
Estimates a reduced rank regression under restrictions.
}
\references{
Hansen, P. R. (2003). Generalized Reduced Rank Regression. Working Paper,
Brown University.
}
