% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstvrrr.R
\name{rstvrrr}
\alias{rstvrrr}
\title{Estimation of RS-TV-RRRs Under Restrictions}
\usage{
rstvrrr(
  yvls,
  xvls,
  uvls,
  rtrs,
  rnrs,
  lrst,
  evar,
  arst = NULL,
  bsbv = FALSE,
  tolr = 100 * sqrt(.Machine$double.eps),
  mxit = Inf,
  mxis = 1
)
}
\arguments{
\item{yvls}{Matrix of dependent variables.}

\item{xvls}{Matrix of (reduced rank coefficients related) regressors.}

\item{uvls}{Matrix of (not reduced rank coefficients related) regressors.}

\item{rtrs}{Specification of the restrictions on the time-varying
row space parameter matrix. Either \code{NULL} or a list. Must have
length 4 if \code{lrst} is \code{NULL}, and length 3 otherwise.
\itemize{
\item If list of length 4, then first element restriction matrix,
second element restriction vector, third element vector of initial values
for the free parameters, and fourth element initial value
for the column space parameter matrix.
\item If list of length 3, then first element restriction matrix,
second element restriction vector, third element vector of initial values
for the free parameters.
}}

\item{rnrs}{Specification of the restrictions on the non-time-varying
row space parameter matrix. Either \code{NULL} or a list. Must have
length 4 only if \code{rtrs} is \code{NULL} and \code{lrst} is \code{NULL};
otherwise must have length 3.
\itemize{
\item If list of length 4, then first element restriction matrix,
second element restriction vector, third element vector of initial values
for the free parameters, and fourth element initial value
for the column space parameter matrix.
\item If list of length 3, then first element restriction matrix,
second element restriction vector, third element vector of initial values
for the free parameters.
}}

\item{lrst}{Specification of the restrictions on the column space
parameter matrix. Either \code{NULL} or a list of length 3. If list of
length 3, then first element restriction matrix, second element restriction
vector, third element vector of initial values for the free parameters.}

\item{evar}{Initial value for the covariance matrix of the error terms
in the measurement equation.}

\item{arst}{Specification of the restrictions on the coefficients of
the input vectors in measurement equation. If not \code{NULL}, \code{NULL} is returned.

Default is \code{NULL}.}

\item{bsbv}{Logical indication whether the state equation considered
by Brune, Scherrer and Bura (2022) should be used.

Default is \code{FALSE}.}

\item{tolr}{Tolerance limit for optimization.

Default is 100 * sqrt(.Machine$double.eps).}

\item{mxit}{Maximum number of iterations (outer loop).

Default is Inf.}

\item{mxis}{Maximum number of iterations (inner loop, switching).

Default is 1.}
}
\value{
Returns \code{NULL} if \code{arst} is not \code{NULL}, otherwise
a list of the following values:
\itemize{
\item \code{kalm}: Results obtained from Kalman smoother
\item \code{sime}: Estimated mean of initial state
\item \code{sicv}: Estimated covariance matrix of initial state
\item \code{evar}: Estimated covariance matrix of measurement equation errors
\item \code{sscf}: Estimated state transition matrix
\item \code{smea}: Estimated drift term in state equation
\item \code{secv}: Estimated covariance matrix of state equation errors
\item \code{augm}: Estimated coefficient matrix of input vectors in measurement
equation
\item \code{lfct}: Estimated column space parameter matrix
\item \code{rnfc}: Estimated non-time-varying part of row space parameter matrix
\item \code{rtfc}: Estimated time-varying part of row space parameter matrix
\item \code{lfur}: Estimated unrestricted column space parameter vector
\item \code{rnur}: Estimated unrestricted non-time-varying row space parameter
vector
\item \code{rtur}: Estimated unrestricted time-varying row space parameter vector
\item \code{nlik}: Negative of the log-likelihood (apart from constants) evaluated
at the optimal solutions found
\item \code{rtrm}: Restriction matrix on time-varying row space parameter matrix
\item \code{rtrv}: Restriction vector on time-varying row space parameter matrix
}
}
\description{
Estimates an RS-TV-RRR under restrictions.
}
\references{
Brune, B., Scherrer, W. and Bura, E. (2022). A State-Space Approach to
Time-Varying Reduced-Rank Regression. Econometric Reviews 41, 895-917.

Veldhuis, S. and Wagner, M. (2026). Restrictions in a State Space Approach
to Time-Varying Vector Error Correction Models: Modelling Instabilities
in Long-Run Money Demand. Mimeo.
}
