% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstvvecm.R
\name{rstvvecm}
\alias{rstvvecm}
\title{Estimation of I(1) RS-TV-VECMs Under Restrictions}
\usage{
rstvvecm(
  tsrs,
  rtrs,
  rnrs,
  lrst,
  evar,
  ordr = 1,
  dpow = 0,
  bsbv = FALSE,
  tolr = 100 * sqrt(.Machine$double.eps),
  mxit = 10000,
  mxis = 1
)
}
\arguments{
\item{tsrs}{Matrix of values for the measured I(1) output.}

\item{rtrs}{Specification of the restrictions on the time-varying
long-run coefficients matrix.  Either \code{NULL} or a list. Must have
length 4 if \code{lrst} is \code{NULL}, and length 3 otherwise.
\itemize{
\item If list of length 4, then first element restriction matrix,
second element restriction vector, third element vector of initial values
for the free parameters, and fourth element initial value
for the adjustment coefficients matrix.
\item If list of length 3, then first element restriction matrix,
second element restriction vector, third element vector of initial values
for the free parameters.
}}

\item{rnrs}{Specification of the restrictions on the non-time-varying
long-run coefficients matrix. Either \code{NULL} or a list. Must have
length 4 only if \code{rtrs} is \code{NULL} and \code{lrst} is \code{NULL};
otherwise must have length 3.
\itemize{
\item If list of length 4, then first element restriction matrix,
second element restriction vector, third element vector of initial values
for the free parameters, and fourth element initial value
for the adjustment coefficients matrix.
\item If list of length 3, then first element restriction matrix,
second element restriction vector, third element vector of initial values
for the free parameters.
}}

\item{lrst}{Specification of the restrictions on the adjustment coefficients
matrix. Either \code{NULL} or a list of length 3. If list of
length 3, then first element restriction matrix, second element restriction
vector, third element vector of initial values for the free parameters.}

\item{evar}{Initial value for the covariance matrix of the error terms
in the measurement equation.}

\item{ordr}{Order of the VECM. Equal to one plus the number of lags.

Default is \code{1}.}

\item{dpow}{Highest power of polynomial time trends (e.g., -1:
no deterministic components, 0: constant only, 1: constant and linear trend)

Default is \code{0}.}

\item{bsbv}{Logical indication whether the state equation considered
by Brune, Scherrer and Bura (2022) should be used.

Default is \code{FALSE}.}

\item{tolr}{Tolerance limit for optimization.

Default is 100 * sqrt(.Machine$double.eps).}

\item{mxit}{Maximum number of iterations (outer loop).

Default is 10000.}

\item{mxis}{Maximum number of iterations (inner loop, switching).

Default is 1.}
}
\value{
Returns \code{NULL} if \code{arst} is not \code{NULL}, otherwise
a list of the following values:
\itemize{
\item \code{kalm}: Results obtained from Kalman smoother
\item \code{sime}: Estimated mean of initial state
\item \code{sicv}: Estimated covariance matrix of initial state
\item \code{evar}: Estimated covariance matrix of measurement equation errors
\item \code{sscf}: Estimated state transition matrix
\item \code{smea}: Estimated drift term in state equation
\item \code{secv}: Estimated covariance matrix of state equation errors
\item \code{augm}: Estimated coefficient matrix of input vectors in measurement
equation
\item \code{lfct}: Estimated adjumstment coefficients matrix
\item \code{rnfc}: Estimated non-time-varying part of long-run coefficients matrix
\item \code{rtfc}: Estimated time-varying part of long-run coefficients matrix
\item \code{lfur}: Estimated unrestricted adjustment coefficients vector
\item \code{rnur}: Estimated unrestricted non-time-varying long-run coefficients
vector
\item \code{rtur}: Estimated unrestricted time-varying long-run coefficients vector
\item \code{nlik}: Negative of the log-likelihood (apart from constants) evaluated
at the optimal solutions found
\item \code{rtrm}: Restriction matrix on time-varying long-run coefficients matrix
\item \code{rtrv}: Restriction vector on time-varying long-run coefficients matrix
\item \code{ordr}: Order of the VECM
\item \code{dpow}: Highest power of polynomial time trends
\item \code{excl}: Logical indicating whether the highest power of polynomial
time trends should only be captured in the long-run relationship
}
}
\description{
Estimates an I(1) RS-TV-VECM under restrictions.
}
\references{
Brune, B., Scherrer, W. and Bura, E. (2022). A State-Space Approach to
Time-Varying Reduced-Rank Regression. Econometric Reviews 41, 895-917.

Veldhuis, S. and Wagner, M. (2026). Restrictions in a State Space Approach
to Time-Varying Vector Error Correction Models: Modelling Instabilities
in Long-Run Money Demand. Mimeo.
}
