% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecm.R
\name{vecm}
\alias{vecm}
\title{Estimation of I(1) VECMs Under Restrictions}
\usage{
vecm(
  tsrs,
  rrst,
  lrst,
  ordr = 1,
  dpow = 0,
  fish = FALSE,
  bima = 0,
  tolr = 100 * sqrt(.Machine$double.eps),
  mxit = 10000
)
}
\arguments{
\item{tsrs}{Matrix of values for multivariate I(1) process.}

\item{rrst}{Specification of the restrictions on the long-run
coefficients matrix. Either \code{NULL} or a list of length 3. If list of
length 3, then first element restriction matrix, second element restriction
vector, third element vector of initial values for the free parameters.}

\item{lrst}{Specification of the restrictions on the adjustment
coefficients matrix. Either \code{NULL} or a list. Must have length 4 if
\code{rrst} is \code{NULL}, and length 3 otherwise.
\itemize{
\item If list of length 4, then first element restriction matrix,
second element restriction vector, third element vector of initial values
for the free parameters, and fourth element initial value
for the long-run coefficients matrix.
\item If list of length 3, then first element restriction matrix,
second element restriction vector, third element vector of initial values
for the free parameters.
}}

\item{ordr}{Order of the VECM. Equal to one plus the number of lags.

Default is \code{1}.}

\item{dpow}{Highest power of polynomial time trends (e.g., -1:
no deterministic components, 0: constant only, 1: constant and linear trend)

Default is \code{0}.}

\item{fish}{Should fisher information matrix be returned?

Default is \code{FALSE}.}

\item{bima}{Order of the Chebyshev time polynomials for a time-varying
VECM as described in Bierens and Martins (2010). Set to 0 for the
standard time-invariant VECM case.

Default is \code{0}.}

\item{tolr}{Tolerance limit for optimization.

Default is 100 * sqrt(.Machine$double.eps).}

\item{mxit}{Maximum number of iterations.

Default is 10000.}
}
\value{
Returns a list of the following values:
\itemize{
\item \code{lfct}: Estimated adjument coefficients matrix
\item \code{rfct}: Estimated long-run coefficients matrix
\item \code{lfun}: Estimated unrestricted adjustment coefficients vector
\item \code{rfun}: Estimated unrestricted long-run coefficients vector
\item \code{evar}: Estimated variance-covariance matrix of the error term
\item \code{augm}: Estimated coefficient matrix of the short-run dynamics
\item \code{lrco}: Cointegrating rank
\item \code{fish}: Fisher information marix (if argument \code{fish} is set to
\code{TRUE}, otherwise \code{NULL})
\item \code{unvr}: Inverse of the Fisher information matrix (if argument
\code{fish} is set to \code{TRUE}, otherwise \code{NULL})
\item \code{idnt}: Logical indicating whether identification is achieved
(if argument \code{fish} is set to \code{TRUE}, otherwise \code{NULL})
\item \code{nlik}: Negative of the log-likelihood (apart from constants) evaluated
at the optimal solutions found
\item \code{ordr}: Order of the VECM
\item \code{dpow}: Highest power of polynomial time trends
\item \code{excl}: Logical indicating whether the highest power of polynomial
time trends should only be captured in the long-run relationship
\item \code{bima}: Order of the Chebyshev time polynomials for a time-varying
VECM as described in Bierens and Martins (2010)
}
}
\description{
Estimates an I(1) VECM under restrictions.
}
\references{
Bierens, H. J. and Martins, L. F. (2010). Time-Varying Cointegration.
Econometric Theory 26, 1453-1490.

Boswijk, H. P. and Doornik, J. A. (2004). Identifying, Estimating and
Testing Restricted Cointegrated Systems: An Overview. Statistica
Neerlandica 58, 440-465.

Doornik, J. A. (1995). Testing General Restrictions on the Cointegrating
Space. Working Paper, Nuffield College, Oxford.

Hansen, P. R. (2003). Generalized Reduced Rank Regression. Working Paper,
Brown University.

Johansen, S. (1995). Likelihood-Based Inference in Cointegrated Vector
Autoregressive Models. Oxford University Press, Oxford.

Pesaran, M. H. and Shin, Y. (2002). Long-Run Structural Modelling.
Econometric Reviews 21, 49-87.
}
